Update address size from 32 to 64 bit
* Update all types that refer to an address from u32 to u64
* Add validation to current targets for values out of range
* Clean up some shared Cortex-M code

+++ debugger/src/debugger/session_data.rs
@@ -24,7 +24,7 @@ pub enum BreakpointType {
 #[derive(Debug)]
 pub struct ActiveBreakpoint {
     pub(crate) breakpoint_type: BreakpointType,
-    pub(crate) breakpoint_address: u32,
+    pub(crate) breakpoint_address: u64,
 }
 
 /// SessionData is designed to be similar to [probe_rs::Session], in as much that it provides handles to the [CoreHandle] instances for each of the available [probe_rs::Core] involved in the debug session.
