+++ probe-rs/src/architecture/arm/component/mod.rs
@@ -152,7 +152,7 @@ fn setup_swv_vendor(
                 // STMicroelectronics:
                 // STM32 parts need TRACE_IOEN set to 1 and TRACE_MODE set to 00.
                 log::debug!("STMicroelectronics part detected, configuring DBGMCU");
-                const DBGMCU: u32 = 0xE004_2004;
+                const DBGMCU: u64 = 0xE004_2004;
                 let mut dbgmcu = memory.read_word_32(DBGMCU)?;
                 dbgmcu |= 1 << 5;
                 dbgmcu &= !(0b00 << 6);
@@ -161,7 +161,7 @@ fn setup_swv_vendor(
             Some(id) if id == jep106::JEP106Code::new(0x02, 0x44) => {
                 // Nordic VLSI ASA
                 log::debug!("Nordic part detected, configuring CLOCK TRACECONFIG");
-                const CLOCK_TRACECONFIG: u32 = 0x4000_055C;
+                const CLOCK_TRACECONFIG: u64 = 0x4000_055C;
                 let mut traceconfig: u32 = 0;
                 traceconfig |= match config.tpiu_clk() {
                     4_000_000 => 3,
